# src/Whisparr.Api.V3/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Whisparr",
    "description": "Whisparr API docs",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/Whisparr/Whisparr/blob/develop/LICENSE"
    },
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "{protocol}://{hostpath}",
      "variables": {
        "protocol": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ]
        },
        "hostpath": {
          "default": "localhost:6969"
        }
      }
    }
  ],
  "paths": {
...
    "/api/v3/episodefile/editor": {
      "put": {
        "tags": [
          "EpisodeFile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeFileListResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v3/episodefile/bulk": {
      "delete": {
        "tags": [
          "EpisodeFile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeFileListResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "EpisodeFile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EpisodeFileResource"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomFormatResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "includeCustomFormatWhenRenaming": {
            "type": "boolean"
          },
          "specifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatSpecificationSchema"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomFormatSpecificationSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "implementation": {
            "type": "string",
            "nullable": true
          },
          "implementationName": {
            "type": "string",
            "nullable": true
          },
          "infoLink": {
            "type": "string",
            "nullable": true
          },
          "negate": {
            "type": "boolean"
          },
          "required": {
            "type": "boolean"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Field"
            },
            "nullable": true
          },
          "presets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatSpecificationSchema"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EpisodeFileListResource": {
        "type": "object",
        "properties": {
          "episodeFileIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "nullable": true
          },
          "quality": {
            "$ref": "#/components/schemas/QualityModel"
          },
          "sceneName": {
            "type": "string",
            "nullable": true
          },
          "releaseGroup": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EpisodeFileResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "seasonNumber": {
            "type": "integer",
            "format": "int32"
          },
          "relativePath": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          },
          "sceneName": {
            "type": "string",
            "nullable": true
          },
          "releaseGroup": {
            "type": "string",
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "nullable": true
          },
          "quality": {
            "$ref": "#/components/schemas/QualityModel"
          },
          "customFormats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatResource"
            },
            "nullable": true
          },
          "mediaInfo": {
            "$ref": "#/components/schemas/MediaInfoResource"
          },
          "qualityCutoffNotMet": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Field": {
        "type": "object",
        "properties": {
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "helpText": {
            "type": "string",
            "nullable": true
          },
          "helpLink": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "advanced": {
            "type": "boolean"
          },
          "selectOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectOption"
            },
            "nullable": true
          },
          "selectOptionsProviderAction": {
            "type": "string",
            "nullable": true
          },
          "section": {
            "type": "string",
            "nullable": true
          },
          "hidden": {
            "type": "string",
            "nullable": true
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacyLevel"
          }
        },
        "additionalProperties": false
      },
...
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MediaInfoResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "audioBitrate": {
            "type": "integer",
            "format": "int64"
          },
          "audioChannels": {
            "type": "number",
            "format": "double"
          },
          "audioCodec": {
            "type": "string",
            "nullable": true
          },
          "audioLanguages": {
            "type": "string",
            "nullable": true
          },
          "audioStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "videoBitDepth": {
            "type": "integer",
            "format": "int32"
          },
          "videoBitrate": {
            "type": "integer",
            "format": "int64"
          },
          "videoCodec": {
            "type": "string",
            "nullable": true
          },
          "videoFps": {
            "type": "number",
            "format": "double"
          },
          "videoDynamicRange": {
            "type": "string",
            "nullable": true
          },
          "videoDynamicRangeType": {
            "type": "string",
            "nullable": true
          },
          "resolution": {
            "type": "string",
            "nullable": true
          },
          "runTime": {
            "type": "string",
            "nullable": true
          },
          "scanType": {
            "type": "string",
            "nullable": true
          },
          "subtitles": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PrivacyLevel": {
        "enum": [
          "normal",
          "password",
          "apiKey",
          "userName"
        ],
        "type": "string"
      },
...
      "Quality": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/QualitySource"
          },
          "resolution": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "QualityModel": {
        "type": "object",
        "properties": {
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          }
        },
        "additionalProperties": false
      },
...
      "QualitySource": {
        "enum": [
          "unknown",
          "television",
          "televisionRaw",
          "web",
          "webRip",
          "dvd",
          "bluray",
          "blurayRaw"
        ],
        "type": "string"
      },
...
      "Revision": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "real": {
            "type": "integer",
            "format": "int32"
          },
          "isRepack": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SelectOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "hint": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "X-Api-Key": {
        "type": "apiKey",
        "description": "Apikey passed as header",
        "name": "X-Api-Key",
        "in": "header"
      },
      "apikey": {
        "type": "apiKey",
        "description": "Apikey passed as header",
        "name": "apikey",
        "in": "query"
      }
    }
  },
  "security": [
    {
      "X-Api-Key": [ ]
    },
    {
      "apikey": [ ]
    }
  ]
}